hardcode sederhana

+++ backend/routes/task.js
@@ -1,12 +1,21 @@
 const express = require("express");
 const router = express.Router();
+
+// const auth = require("../middleware/auth");
 const db = require("../db");
 
 /**
  * GET all tasks
  */
 router.get("/", (req, res) => {
+// router.get("/", auth, (req, resc) => {
+    const apiKey = req.headers["x-api-key"]
+
+  if (apiKey !== "1234567") {
+    return res.status(401).json({ message: "Unauthorized" })
+  }
   db.query("SELECT * FROM tasks", (err, result) => {
+    console.log(result);
     if (err) return res.status(500).json(err);
     res.json(result);
   });
@@ -16,6 +25,11 @@ router.get("/", (req, res) => {
  * GET task by id
  */
 router.get("/:id", (req, res) => {
+    const apiKey = req.headers["x-api-key"]
+
+  if (apiKey !== "1234567") {
+    return res.status(401).json({ message: "Unauthorized" })
+  }
   db.query(
     "SELECT * FROM tasks WHERE id = ?",
     [req.params.id],
@@ -30,7 +44,16 @@ router.get("/:id", (req, res) => {
  * CREATE task
  */
 router.post("/", (req, res) => {
+  const apiKey = req.headers["x-api-key"]
+  
+  
+  if (apiKey !== "1234567") {
+    return res.status(401).json({ message: "Unauthorized" })
+  }
   const { title, description, status, due_date } = req.body;
+   if (!title) {
+    return res.status(400).json({ message: "Title is required" })
+  }
 
   const sql =
     "INSERT INTO tasks (title, description, status, due_date) VALUES (?,?,?,?)";
@@ -45,6 +68,11 @@ router.post("/", (req, res) => {
  * UPDATE task
  */
 router.put("/:id", (req, res) => {
+  const apiKey = req.headers["x-api-key"]
+
+  if (apiKey !== "1234567") {
+    return res.status(401).json({ message: "Unauthorized" })
+  }
   const { title, description, status, due_date } = req.body;
 
   const sql =
@@ -64,6 +92,11 @@ router.put("/:id", (req, res) => {
  * DELETE task
  */
 router.delete("/:id", (req, res) => {
+  const apiKey = req.headers["x-api-key"]
+
+  if (apiKey !== "1234567") {
+    return res.status(401).json({ message: "Unauthorized" })
+  }
   db.query(
     "DELETE FROM tasks WHERE id=?",
     [req.params.id],
